requirements.txt
openai-whisper
pydub
librosa
sentence-transformers
chromadb
pyannote.audio
torch
streamlit
